+++ lib/aggregate/attribute/bitfield.rb
@@ -40,10 +40,6 @@ class Aggregate::Attribute::Bitfield < Aggregate::Attribute::Base
     []
   end
 
-  def skip_default?
-    true
-  end
-
   private
 
   def value_to_string(value)