+++ test/dummy/app/models/passport.rb
@@ -9,7 +9,7 @@ class Passport < ActiveRecord::Base
   aggregate_attribute :state,            :string,   required: true
   aggregate_attribute :birthdate,        :datetime, required: true
   aggregate_attribute :height,           :decimal
-  aggregate_attribute :weight,           :decimal, default: 100
+  aggregate_attribute :weight,           :decimal, default: 100, write_default_values: true
   aggregate_attribute :photo,            "PassportPhoto"
   aggregate_has_many  :foreign_visits,   "ForeignVisit"
   aggregate_attribute :stamps,           :bitfield, limit: 10